move filter option types
