--- frontend/components/ResultsPage/filterTypes.ts
@@ -1,10 +1,0 @@
-export type Option = {
-  value: string,
-  count: number
-}
-
-export type FilterOptions = {
-  nupath: Option[],
-  subject: Option[],
-  classType: Option[],
-}
