+++ frontend/components/types.ts
@@ -5,9 +5,6 @@
  * ONLY PUT COMMONLY USED TYPES HERE
  */
 
-import { FilterOptions } from './ResultsPage/filterTypes';
-
-
 // ======= Search Results ========
 // Represents the course and employee data returned by /search
 export interface SearchResult {
@@ -31,4 +28,17 @@ export interface FilterSelection {
   nupath?: string[],
   subject?: string[],
   classType?: string[],
+}
+
+// Represents an option in a dropdown or checkbox filter
+export type Option = {
+  value: string,
+  count: number
+}
+
+// represents the options for all filters
+export type FilterOptions = {
+  nupath: Option[],
+  subject: Option[],
+  classType: Option[],
 }
